Bug fix: deselection doesn't work properly
In the current code, deselection doesn't work as expected, since you need tagList items and multiselect value items to be in the same order. 
Otherwise, you'll receive the e.item and e.dataItem parameters of the deselect event referencing to different items. And the deselection itself occurs in the order you've been adding the items, not the sorted one. To prevent this, I propose to update the value of multiselect, ordering the values the same way you order tags. Hope this will help you